+++ scripts/config.sh
@@ -1,0 +1,1 @@
+MAPPING_URL=https://raw.githubusercontent.com/sfu-ireceptor/config/turnkey-v4/AIRR-iReceptorMapping.txt