+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Release Notes
 
+## v4.0-pre (Sep 9, 2022)
+- Enable [customization of the home page](doc/customizing_home_page.md)
+
+## v4.0-pre (June 13, 2022)
+- Enable [customization of the /airr/v1/info entry point](doc/customizing_info_entry_point.md)
+
+## v4.0-pre (May 19, 2022)
+- Use systemd to start turnkey on boot
+- Minor bug fixes
+
+## v4.0-pre (Apr 4, 2022)
+- Add clone support
+- Add cell support
+- Add expression support
+- Add stats removal
+
+## v4.0-pre (Dec 22, 2021)
+- Use HTTPS by default (with a generated self-signed SSL certificate)
+- Use web service ``turnkey-v4`` branch
+- Add stats loading script
+- Add clone loading script
+- Add metadata update script
+
 ## v3.1 (May 10, 2021)
 
 ### Web service update